+++ tests/upstream_alignment_test.py
@@ -91,6 +91,7 @@ TEST_IDS = (
     'missing-whitespace-before-tag',
     'mjml-comment-merging',
     'shorthand-color-attributes',
+    'xml-boolean-attributes',
 )
 @pytest.mark.parametrize('test_id', TEST_IDS)
 def test_ensure_same_html_as_upstream(test_id, fixed_random_seed):